reports: persist every result row and consolidate result storage
The row cap ran in finishSuccess, after the script had already built the full
result in memory, so it could not bound execution memory; results are stored as
a gzipped binary anyway, with no JCR property-size limit. Drop capRows, the
truncated flag (execution + preview + JSON + read) and the now-unused
maxResultRows OSGi property, so every row is persisted.

Move the result-write logic into ReportResultStore so it owns both read and
write, and format dates with java.time instead of SimpleDateFormat.
