+++ extensions/reports/api/src/main/groovy/be/orbinson/aem/groovy/console/reports/model/ReportPreview.groovy
@@ -14,11 +14,8 @@ class ReportPreview {
     /** SUCCESS or FAILED. */
     ReportExecutionStatus status
 
-    /** Typed result (columns + capped rows); null when the run failed. */
+    /** Typed result (columns + rows); null when the run failed. */
     ReportData data
-
-    /** Whether the result was capped to the configured maximum row count. */
-    boolean truncated
 
     /** Captured script output (println etc.), possibly truncated. */
     String output
