+++ extensions/reports/bundle/src/main/groovy/be/orbinson/aem/groovy/console/reports/impl/ReportsConfigurationService.groovy
@@ -7,7 +7,5 @@ interface ReportsConfigurationService {
 
     int getDefaultPageSize()
 
-    int getMaxResultRows()
-
     int getMaxOutputLength()
 }
