+++ extensions/reports/bundle/src/main/java/be/orbinson/aem/groovy/console/reports/impl/ReportsConfigurationProperties.java
@@ -12,10 +12,6 @@ public @interface ReportsConfigurationProperties {
         description = "Default result page size in the reports UI when a report does not declare its own.")
     int defaultPageSize() default 50;
 
-    @AttributeDefinition(name = "Maximum Result Rows",
-        description = "Maximum number of result rows persisted per execution.  Results exceeding this limit are truncated and flagged.")
-    int maxResultRows() default 10000;
-
     @AttributeDefinition(name = "Maximum Output Length",
         description = "Maximum number of characters of captured script output persisted per execution.")
     int maxOutputLength() default 100000;
