+++ it.tests/src/main/features/groovyconsole-reports.json
@@ -12,10 +12,5 @@
       "id": "org.apache.commons:commons-compress:1.24.0",
       "start-order": "19"
     }
-  ],
-  "configurations":{
-    "be.orbinson.aem.groovy.console.reports.impl.DefaultReportsConfigurationService":{
-      "maxResultRows:Integer":5
-    }
-  }
+  ]
 }
